коммит для 14-го спринта

+++ controllers/cards.js
@@ -1,34 +1,46 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/NotFoundError');
+const NoRightsError = require('../errors/NoRightsError');
+const InaccurateDataError = require('../errors/InaccurateDataError');
 
 // получить все карточки
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(500).send({ message: 'Ошибка сервера' }));
+    .catch(next);
 };
 
 // удалить карточку
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
-    .then((deletedCard) => {
-      if (!deletedCard) {
-        res.status(404).send({ message: 'Карточка не найдена' });
-        return;
+  const userId = req.user.id;
+
+  Card.findById(cardId)
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Карточка не найдена');
       }
-      res.status(200).send(deletedCard);
+
+      if (card.owner !== userId) {
+        throw new NoRightsError('Нет прав на удаление карточки');
+      }
+
+      Card.findByIdAndRemove(cardId)
+        .then((deletedCard) => {
+          res.status(200).send(deletedCard);
+        })
+        .catch(next);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Передан неверный id карточки' });
-        return;
+        throw new InaccurateDataError('Передан неверный id карточки');
       }
-      res.status(500).send({ message: 'Ошибка сервера' });
+      next(err);
     });
 };
 
 // создать карточку
-module.exports.postCard = (req, res) => {
+module.exports.postCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -38,51 +50,48 @@ module.exports.postCard = (req, res) => {
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(400).send({ message: 'Переданы некорректные данные' });
-        return;
+        throw new InaccurateDataError('Переданы некорректные данные');
       }
-      res.status(500).send({ message: 'Ошибка сервера' });
+      next(err);
     });
 };
 
 // лайкнуть карточку
-module.exports.likeCard = (req, res) => Card.findByIdAndUpdate(
+module.exports.likeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $addToSet: { likes: req.user._id } },
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
-      return;
+      throw new NotFoundError('Запрашиваемая карточка не найдена');
     }
     res.send(card);
   })
   .catch((err) => {
     if (err.name === 'CastError') {
-      res.status(400).send({ message: 'Был передан несуществующий id карточки' });
+      throw new InaccurateDataError('Был передан несуществующий id карточки');
     } else {
-      res.status(500).send({ message: 'Ошибка сервера' });
+      next(err);
     }
   });
 
 // убрать лайк карточки
-module.exports.dislikeCard = (req, res) => Card.findByIdAndUpdate(
+module.exports.dislikeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $pull: { likes: req.user._id } },
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
-      return;
+      throw new NotFoundError('Запрашиваемая карточка не найдена');
     }
     res.send(card);
   })
   .catch((err) => {
     if (err.name === 'CastError') {
-      res.status(400).send({ message: 'Был передан несуществующий id карточки' });
+      throw new InaccurateDataError('Был передан несуществующий id карточки');
     } else {
-      res.status(500).send({ message: 'Ошибка сервера' });
+      next(err);
     }
   });
